requirements.txt
Django>=3.2.3
django-crispy-forms >= 1.11.2
pillow >= 5.3.0